+++ emsim/ext/src/dens_kernel.h
@@ -2,7 +2,7 @@
 #define DENS_KERNEL_H_
 
 int build_slices_fftwf_kernel(float scattering_factors_ifftshifted[], int n_elems,
-                             unsigned int atom_histograms[], int n_slices, int len_x, int len_y,
-                             float output[]);
+                              float atom_histograms[], int n_slices, int len_x, int len_y,
+                              float output[]);
 
 #endif
